common-aliases: Make media file viewer configurable using the `$XMVIEWER` variable

+++ plugins/common-aliases/common-aliases.plugin.zsh
@@ -120,8 +120,10 @@ if is-at-least 4.2.0; then
     for ft in $_image_fts; do alias -s $ft='$XIVIEWER'; done
   fi
 
-  _media_fts=(ape avi flv m4a mkv mov mp3 mpeg mpg ogg ogm rm wav webm)
-  for ft in $_media_fts; do alias -s $ft=mplayer; done
+  if [[ -n "$XMVIEWER" ]]; then
+    _media_fts=(ape avi flv m4a mkv mov mp3 mpeg mpg ogg ogm rm wav webm)
+    for ft in $_media_fts; do alias -s $ft=$XMVIEWER; done
+  fi
 
   #read documents
   alias -s pdf=acroread
